+++ CHANGELOG.md
@@ -7,4 +7,5 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added provider configuration.
+- Added organizations data-source.
 - Added users data-source.